requirements.txt
boto3>=1.34.0
PyJWT>=2.8.0
yt-dlp>=2024.1.0
